Moving flavor_transformation to a separate dir
